tmf: Add recyclable attributes for state system analyses
This new class allows to reuse attribute quarks in a state system. It can
typically be used by analyses who want to save data in the state system
for short intervals of time. Instead of creating a number of attributes with
lots of empty space before and after, they can reuse the same. It makes for
more compact time graph views as well.

Change-Id: Ie7f32eccb02096ef62df4c479a8e2bcfcd043af6
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>